Some updates to hit display and strip charts
My hit display now contains code for selecting a specific spill to look at. This remains unfinished as the looping to find new files continues, but I am putting it on hold until Jay changes the hit display to be on the event level.

What were previously known as "SpillCharts.py" and "Stripcharts.py" have now been merged into one file in "StripCharts.py". This was done in order to control both spill windows with a single button.