Created MassHist.py
Created MassHist.py which shows a histogram of all masses for events in a given spill. Functionalities include automatically searching for new data, manually selecting a spill to view, and returning to the most recent spill.

Also made some small changes to App.py and DataReader.py to run MassHist.py

One odd thing to note is that Arthur's code (the calcVariables function) returns 6 masses for each event, I dont know enough to say if it should be doing this but like I said, it seems odd.